update unit test.
